fea: ex_equip_recycle
